+++ demos/chinook/src/main/java/is/codion/demos/chinook/model/InvoiceModel.java
@@ -18,6 +18,7 @@
  */
 package is.codion.demos.chinook.model;
 
+import is.codion.demos.chinook.domain.api.Chinook.InvoiceLine;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.model.ForeignKeyDetailModelLink;
 import is.codion.swing.framework.model.SwingEntityModel;
@@ -32,14 +33,15 @@ public final class InvoiceModel extends SwingEntityModel {
 		InvoiceLineEditModel invoiceLineEditModel = new InvoiceLineEditModel(connectionProvider);
 
 		SwingEntityModel invoiceLineModel = new SwingEntityModel(invoiceLineEditModel);
-		ForeignKeyDetailModelLink<?, ?, ?> detailModelLink = detailModels().add(invoiceLineModel);
-		// Prevents accidentally adding a new invoice line to the previously selected invoice,
-		// since the selected foreign key value persists when the master selection is cleared by default.
-		detailModelLink.clearValueOnEmptySelection().set(true);
-		// Usually the UI is responsible for activating the detail model link for the currently
-		// active (or visible) detail panel, but since the InvoiceLine panel is embedded in the
-		// InvoiceEditPanel, we simply activate the link here.
-		detailModelLink.active().set(true);
+		detailModels().add(ForeignKeyDetailModelLink.builder(invoiceLineModel, InvoiceLine.INVOICE_FK)
+						// Prevents accidentally adding a new invoice line to the previously selected invoice,
+						// since the selected foreign key value persists when the master selection is cleared by default.
+						.clearValueOnEmptySelection(true)
+						// Usually the UI is responsible for activating the detail model link for the currently
+						// active (or visible) detail panel, but since the InvoiceLine panel is embedded in the
+						// InvoiceEditPanel, we simply activate the link here.
+						.active(true)
+						.build());
 
 		// We listen for when invoice totals are updated by the edit model,
 		// and replace the invoices in the table model with the updated ones.