+++ framework/model/src/test/java/is/codion/framework/model/DefaultEntityModelTest.java
@@ -35,7 +35,9 @@ public class DefaultEntityModelTest extends AbstractEntityModelTest<DefaultEntit
 	protected TestEntityModel createDepartmentModel() {
 		TestEntityModel deptModel = new TestEntityModel(new TestEntityEditModel(Department.TYPE, connectionProvider()));
 		TestEntityModel empModel = new TestEntityModel(new TestEntityEditModel(Employee.TYPE, connectionProvider()));
-		deptModel.detailModels().add(empModel).active().set(true);
+		deptModel.detailModels().add(DetailModelLink.builder(empModel)
+						.active(true)
+						.build());
 
 		return deptModel;
 	}